Renamed grammar files to .langium, moved generated code to subfolder
