+++ cli/src/package.ts
@@ -7,8 +7,7 @@ export interface Package {
 
 export interface LangiumConfig {
     grammar?: string,
-    extension?: string,
-    name?: string,
+    extensions?: string[],
     out?: string,
     // The following option is meant to be used only by Langium itself
     langiumInternal?: boolean
