+++ core/package.json
@@ -9,15 +9,16 @@
     "build": "tsc --build tsconfig.build.json",
     "watch": "tsc --build tsconfig.build.json --watch",
     "lint": "eslint ./src --ext .js,.ts",
-    "prepare": "npm run clean && npm run lint && npm run build && copyfiles -f ./src/gen/grammar.json ./lib/gen",
+    "prepare": "npm run clean && npm run lint && npm run build && copyfiles -f ./src/grammar-lang/generated/grammar.json ./lib/grammar-lang/generated",
     "bootstrap": "node ../cli/lib/langium",
     "test": "jest --coverage"
   },
   "author": "",
   "license": "MIT",
   "langium": {
-    "grammar": "./langium.lg",
-    "extension": ".lg",
+    "grammar": "src/grammar-lang/langium-grammar.langium",
+    "extensions": ["langium"],
+    "out": "src/grammar-lang/generated",
     "langiumInternal": true
   },
   "dependencies": {
