+++ core/src/generator/utils.ts
@@ -1,4 +1,4 @@
-import * as ast from '../gen/ast';
+import * as ast from '../grammar-lang/generated/ast';
 import { replaceTokens } from './token-replacer';
 
 type FeatureValue = {
