+++ core/src/grammar-lang/langium-grammar.langium
@@ -1,4 +1,4 @@
-grammar Langium hidden(WS)
+grammar LangiumGrammar hidden(WS)
 
 terminal WS         : /\s+/;
 terminal ID returns string: /\^?[_a-zA-Z][\w_]*/;
@@ -9,8 +9,7 @@ Grammar:
 	'grammar' name=ID ('with' usedGrammars+=[Grammar|ID] (',' usedGrammars+=[Grammar|ID])*)?
 	(definesHiddenTokens?='hidden' '(' (hiddenTokens+=[AbstractRule|ID] (',' hiddenTokens+=[AbstractRule|ID])*)? ')')?
 	metamodelDeclarations+=AbstractMetamodelDeclaration*
-	(rules+=AbstractRule)+
-;
+	(rules+=AbstractRule)+;
 
 AbstractRule : ParserRule | TerminalRule | EnumRule;
 
@@ -25,8 +24,7 @@ ReferencedMetamodel :
 
 
 Annotation:
-	'@' name = ID
-;
+	'@' name = ID;
 
 ParserRule :
 	(
@@ -35,41 +33,32 @@ ParserRule :
 	)
 	(definesHiddenTokens?='hidden' '(' (hiddenTokens+=[AbstractRule|ID] (',' hiddenTokens+=[AbstractRule|ID])*)? ')')? ':'
 		alternatives=Alternatives
-	';'
-;
+	';';
 
 fragment RuleNameAndParams:
-	name=ID ('<' (parameters+=Parameter (',' parameters+=Parameter)*)? '>')?
-;
+	name=ID ('<' (parameters+=Parameter (',' parameters+=Parameter)*)? '>')?;
 
 Parameter:
-	name=ID
-;
+	name=ID;
 
 Alternatives returns AbstractElement:
-	UnorderedGroup ({Alternatives.elements+=current} '|' elements+=UnorderedGroup)*
-;
+	UnorderedGroup ({Alternatives.elements+=current} '|' elements+=UnorderedGroup)*;
 
 UnorderedGroup returns AbstractElement:
-	Group ({UnorderedGroup.elements+=current} '&' elements+=Group)*
-;
+	Group ({UnorderedGroup.elements+=current} '&' elements+=Group)*;
 
 Group returns AbstractElement:
-	AbstractToken ({Group.elements+=current} elements+=AbstractToken)*
-;
+	AbstractToken ({Group.elements+=current} elements+=AbstractToken)*;
 
 AbstractToken returns AbstractElement:
 	AbstractTokenWithCardinality |
-	Action
-;
+	Action;
 
 AbstractTokenWithCardinality returns AbstractElement:
-	(Assignment | AbstractTerminal) cardinality=('?'|'*'|'+')?
-;
+	(Assignment | AbstractTerminal) cardinality=('?'|'*'|'+')?;
 
 Action returns AbstractElement:
-	{Action} '{' type=ID ('.' feature=ID operator=('='|'+=') 'current')? '}'
-;
+	{Action} '{' type=ID ('.' feature=ID operator=('='|'+=') 'current')? '}';
 
 AbstractTerminal returns AbstractElement:
 	Keyword |
@@ -77,152 +66,116 @@ AbstractTerminal returns AbstractElement:
 	ParenthesizedElement |
 	PredicatedKeyword |
 	PredicatedRuleCall |
-	PredicatedGroup
-;
+	PredicatedGroup;
 
 Keyword:
-	value=string
-;
+	value=string;
 
 RuleCall:
-	rule=[AbstractRule|ID] ('<' arguments+=NamedArgument (',' arguments+=NamedArgument)* '>')?
-;
+	rule=[AbstractRule|ID] ('<' arguments+=NamedArgument (',' arguments+=NamedArgument)* '>')?;
 
 NamedArgument:
 	( parameter=[Parameter|ID] calledByName?= '=')?
-	( value=Disjunction )
-;
+	( value=Disjunction );
 
 LiteralCondition:
-	^true?='true' | 'false'
-;
+	^true?='true' | 'false';
 
 Disjunction returns Condition:
-	Conjunction ({Disjunction.left=current} '|' right=Conjunction)?
-;
+	Conjunction ({Disjunction.left=current} '|' right=Conjunction)?;
 
 Conjunction returns Condition:
-	Negation ({Conjunction.left=current} '&' right=Negation)?
-;
+	Negation ({Conjunction.left=current} '&' right=Negation)?;
 
 Negation returns Condition:
-	Atom | {Negation} '!' value=Negation
-;
+	Atom | {Negation} '!' value=Negation;
 
 Atom returns Condition:
-	ParameterReference | ParenthesizedCondition | LiteralCondition
-;
+	ParameterReference | ParenthesizedCondition | LiteralCondition;
 
 ParenthesizedCondition returns Condition:
-	'(' Disjunction ')'
-;
+	'(' Disjunction ')';
 
 ParameterReference:
-	parameter=[Parameter|ID]
-;
+	parameter=[Parameter|ID];
 
 TerminalRuleCall:
-	rule=[AbstractRule|ID]
-;
+	rule=[AbstractRule|ID];
 
 PredicatedKeyword returns Keyword:
-	(predicated?='=>' | firstSetPredicated?='->') value=string
-;
+	(predicated?='=>' | firstSetPredicated?='->') value=string;
 
 PredicatedRuleCall returns RuleCall:
-	(predicated?='=>' | firstSetPredicated?='->') rule=[AbstractRule|ID] ('<' arguments+=NamedArgument (',' arguments+=NamedArgument)* '>')?
-;
+	(predicated?='=>' | firstSetPredicated?='->') rule=[AbstractRule|ID] ('<' arguments+=NamedArgument (',' arguments+=NamedArgument)* '>')?;
 
 Assignment returns AbstractElement:
-	{Assignment} (predicated?='=>' | firstSetPredicated?='->')? feature=ID operator=('+='|'='|'?=') ^terminal=AssignableTerminal
-;
+	{Assignment} (predicated?='=>' | firstSetPredicated?='->')? feature=ID operator=('+='|'='|'?=') ^terminal=AssignableTerminal;
 
 AssignableTerminal returns AbstractElement:
-	Keyword | RuleCall | ParenthesizedAssignableElement | CrossReference
-;
+	Keyword | RuleCall | ParenthesizedAssignableElement | CrossReference;
 
 ParenthesizedAssignableElement returns AbstractElement:
-	'(' AssignableAlternatives ')'
-;
+	'(' AssignableAlternatives ')';
 
 AssignableAlternatives returns AbstractElement:
-	AssignableTerminal ({Alternatives.elements+=current} ('|' elements+=AssignableTerminal)+)?
-;
+	AssignableTerminal ({Alternatives.elements+=current} ('|' elements+=AssignableTerminal)+)?;
 
 CrossReference returns AbstractElement:
-	{CrossReference} '[' type=[ParserRule|ID] ('|' ^terminal=CrossReferenceableTerminal )? ']'
-;
+	{CrossReference} '[' type=[ParserRule|ID] ('|' ^terminal=CrossReferenceableTerminal )? ']';
 
 CrossReferenceableTerminal returns AbstractElement:
-	Keyword | RuleCall
-;
+	Keyword | RuleCall;
 
 ParenthesizedElement returns AbstractElement:
-	'(' Alternatives ')'
-;
+	'(' Alternatives ')';
 
 PredicatedGroup returns Group:
-	(predicated?='=>' | firstSetPredicated?='->') '(' elements+=Alternatives ')'
-;
+	(predicated?='=>' | firstSetPredicated?='->') '(' elements+=Alternatives ')';
 
 TerminalRule :
 	'terminal' (^fragment?='fragment' name=ID | name=ID ('returns' type=ID)?) ':'
 		regex=RegexLiteral
-	';'
-;
+	';';
 
 terminal RegexLiteral returns string: /\/(?![*+?])(?:[^\r\n\[/\\]|\\.|\[(?:[^\r\n\]\\]|\\.)*\])+\//;
 
 TerminalAlternatives:
-	TerminalGroup ({TerminalAlternatives.elements+=current} '|' elements+=TerminalGroup)*
-;
+	TerminalGroup ({TerminalAlternatives.elements+=current} '|' elements+=TerminalGroup)*;
 
 TerminalGroup:
-	elements+=TerminalToken*
-;
+	elements+=TerminalToken*;
 
 TerminalToken:
-	TerminalTokenElement cardinality=('?'|'*'|'+')?
-;
+	TerminalTokenElement cardinality=('?'|'*'|'+')?;
 
 TerminalTokenElement:
-	CharacterRange | TerminalRuleCall | ParenthesizedTerminalElement | AbstractNegatedToken | Wildcard
-;
+	CharacterRange | TerminalRuleCall | ParenthesizedTerminalElement | AbstractNegatedToken | Wildcard;
 
 ParenthesizedTerminalElement returns TerminalAlternatives:
-	'(' TerminalAlternatives ')'
-;
+	'(' TerminalAlternatives ')';
 
 AbstractNegatedToken:
-	NegatedToken | UntilToken
-;
+	NegatedToken | UntilToken;
 
 NegatedToken:
-	'!' ^terminal=TerminalTokenElement
-;
+	'!' ^terminal=TerminalTokenElement;
 
 UntilToken:
-	'->' ^terminal=TerminalTokenElement
-;
+	'->' ^terminal=TerminalTokenElement;
 
 Wildcard:
-	{Wildcard} '.'
-;
+	{Wildcard} '.';
 
 CharacterRange:
-	left=Keyword ('..' right=Keyword)?
-;
+	left=Keyword ('..' right=Keyword)?;
 
 EnumRule:
 	'enum' name=ID ('returns' type=ID)? ':'
 		alternatives=EnumLiterals
-	';'
-;
+	';';
 
 EnumLiterals:
-	EnumLiteralDeclaration ({EnumLiterals.elements+=current} '|' elements+=EnumLiteralDeclaration)*
-;
+	EnumLiteralDeclaration ({EnumLiterals.elements+=current} '|' elements+=EnumLiteralDeclaration)*;
 
 EnumLiteralDeclaration:
-	enumLiteral=ID ('=' literal=Keyword)?
-;
+	enumLiteral=ID ('=' literal=Keyword)?;
