+++ core/src/index.ts
@@ -1,16 +1,16 @@
 export * from './dependency-injection';
 export * from './default-module';
 export * from './services';
-export * from './gen/ast';
-export * from './gen/grammar-access';
-export * from './gen/module';
-export * from './gen/parser';
 export * from './generator/ast-node';
 export * from './generator/ast-util';
 export * from './generator/node';
 export * from './generator/node-processor';
 export * from './generator/template-string';
 export * from './generator/token-replacer';
+export * from './grammar-lang/generated/ast';
+export * from './grammar-lang/generated/grammar-access';
+export * from './grammar-lang/generated/module';
+export * from './grammar-lang/generated/parser';
 export * from './service/validation/document-validator';
 export * from './service/validation/validation-registry';
 export * from './generator/utils';
